+++ builder-core/src/attribute/sub/offensive.rs
