+++ builder-core/src/attribute/sub/spell_school.rs
@@ -1,4 +1,4 @@
-use crate::{logic::attribute::Attribute, simple_enum};
+use crate::{attribute::Attribute, simple_enum};
 
 simple_enum!(SpellSchool, (Abjuration "Abjuration", Conjuration "Conjuration", Divination "Divintation", Enchantment "Enchantment", Evocation "Evocation", Illusion "Illusion", Necromancy "Necromancy", Transmutation "Transmutation", All "All"));
