+++ builder-core/src/attribute/sub/weapon.rs
