+++ builder-core/src/lib.rs
@@ -1,5 +1,7 @@
+#![allow(dead_code)]
 pub mod attribute;
 pub mod bonus;
 pub mod breakdown;
 pub mod character;
 pub mod feat;
+pub mod utils;
