+++ builder-core/src/utils/simple_enum.rs
